Send message with invalid replyto queue to error queue
When replying to a synchronous message, currently if the message cannot be sent because of an invalid replyto queue it is retried 10 times and then sent to the deadletter queue.

Retrying sending is not useful for a synchronous message that will timeout before it can be retried. Resending will also not work since the queue is likely not there later. Instead we send the message to the error queue of the sender to inform of the error with an error message transport:invalid-field-value.

Also added support for sending a message to the deadletter queue immediately. This is currently unused.

PR #42