ServiceBusFactoryPool: Do not increment ActiveCount beyond 1
The ServiceBusFactoryPool issues ServiceBusFactories which contains the
actual AMQP Connection which should not be recycled unless we shut down
or are disconnected because of network issues.

Because of this Release will never be called for a ServiceBusFactory and
ActiveCount would just be incremented forever until restart.

This commit let's us override this by setting the member variable
_incrementActiveCount of MessagingEntityCache<T> to false after creation
of the necessary ServiceBusFactories. With this change ActiveCount for a
ServiceBusFactory will never exceed ActiveCount of 1.