Examples: Add example w/decryption and signature validation capabilities
This example showcases how to use the SignThenEncryptMessageProtection
as well as illustrating a custom implementations of ICertificateStore
and IAddressRegistry.

ICertificateStore is useful if your certificates are stored elsewhere
than the Windows Certificate Store or you require a mock like in this
example.

The IAddressRegistry implementation is also use when you require a mock.
In most real-world application you would use the AddressRegistry
implementation of IAddressRegistry.