MessagingServer: Do not Ping() the CPPAService
The Ping method for the CPPAService tries to download the full
CollaborationProtocolProfile. We do not have proper serialization of
certificates for CollaborationProtcolProfile like we do for
CollaborationProtocolAgreement so let us not call this method for now.

In the future the Ping method for the CPPAService will be a proper Ping
implementation.