Tests+Examples: Use self-signed certificates
This adds convencience methods for generating self-signed certificates
and implements usage of self-signed certificates in some of the Examples
and for our tests in Helsenorge.Messaging.Tests.

The example applications Helsenorge.Messaging.Client and
Helsenorge.Messaging.Server uses certificates from the Windows
Certificate Store and is meant to be as close to a real-life application
as possible.

Helsenorge.Registries.Tests does not yet use self-signed certificates,
but will do so in a future PR.

Co-authored-by: Bjarte Lund <bjarte@godkjemi.net>