Messaging: Don't call EnsureOpen before doing operations on AMQP message
This patch removes the EnsureOpen{,Async} calls before doing an
operation on the AMQP message.

If we recreate the link before executing operations on the already
retrieved message we will just get an exception telling us that the
message is not associated with the link, so no point trying to recreate
the link at this point.