Registries: Add support for old CPA/CPP Cache
Add support for old CollaborationProtocolProfile cached objects.

The new property Action becomes null when reading from old Cache, as the
old cache was created without this property. Action in the new version
maps from the same value as Name did in the old version so the fix is
adding a fallback so if Action is null then we'll use the Name property
instead. This fallback does also work with new caches where Action is
being cached.