Tests+Examples: Move mocks into a separate project
To be able to upgrade to MSTest.TestAdapter version 3.2.0, and since we
are using the mocks in some of our example projects I had to move the
mocks into a separate library to avoid the error: "CS0017 Program has
more than one entry point defined. Compile with /main to specify the
type that contains the entry point."

Not sure why this didn't hit us before, but at least this approach fixes
the problem.

For more information see: https://github.com/xunit/xunit/issues/1172